Helper utils tests
