--- pkg/jobs/create/hive.go
@@ -1,52 +1,0 @@
-// Copyright Contributors to the Open Cluster Management project.
-package create
-
-import (
-	"context"
-	"encoding/json"
-	"errors"
-	"log"
-
-	hiveclient "github.com/openshift/hive/pkg/client/clientset/versioned"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/klog/v2"
-)
-
-//  patchStringValue specifies a json patch operation for a string.
-type patchStringValue struct {
-	Op    string `json:"op"`
-	Path  string `json:"path"`
-	Value int32  `json:"value"`
-}
-
-func ActivateDeploy(hiveset hiveclient.Interface, clusterName string) error {
-	klog.V(0).Info("* Initiate Provisioning")
-	klog.V(2).Info("Looking up cluster " + clusterName)
-	cluster, err := hiveset.HiveV1().ClusterDeployments(clusterName).Get(context.TODO(), clusterName, v1.GetOptions{})
-	if err != nil {
-		return err
-	}
-
-	klog.V(2).Info("Found cluster " + cluster.Name + " ✓")
-	if cluster.Spec.InstallAttemptsLimit == nil || *cluster.Spec.InstallAttemptsLimit != 0 {
-		return errors.New("ClusterDeployment.spec.installAttemptsLimit is not 0")
-	}
-
-	// Update the installAttemptsLimit
-	intValue := int32(1)
-	patch := []patchStringValue{{
-		Op:    "replace",
-		Path:  "/spec/installAttemptsLimit",
-		Value: intValue,
-	}}
-	patchInBytes, _ := json.Marshal(patch)
-	*cluster.Spec.InstallAttemptsLimit = 1
-	_, err = hiveset.HiveV1().ClusterDeployments(clusterName).Patch(
-		context.TODO(), clusterName, types.JSONPatchType, patchInBytes, v1.PatchOptions{})
-	if err != nil {
-		return err
-	}
-	log.Println("Updated ClusterDeployment ✓")
-	return nil
-}
